[ABT-236] Refactor callback pattern to favor exceptions.
Exceptions provide a more intuitive interface into the underlying
reason why the SDK failed to init/load/run. The callbacks were used
synchronously so they were of little use to have the code in a seperate
class altogether. If we need to investigate having the SDK load
asynchronously, then we'll investigate Promises and Futures (through
AsyncTask on Android).