Unity SDK 2.1.0 support and backwards compatibility.
* Register a single event listener with Unity SDK for all placements and route to multiple sub listeners based on placement id

* [ABT-236] Refactor callback pattern to favor exceptions.

Exceptions provide a more intuitive interface into the underlying
reason why the SDK failed to init/load/run. The callbacks were used
synchronously so they were of little use to have the code in a seperate
class altogether. If we need to investigate having the SDK load
asynchronously, then we'll investigate Promises and Futures (through
AsyncTask on Android).

* Delete exteraneous files and move them to the UnityRouter class

* Comment out the @Override annotation for the time being.

Ensuring that this method won't cause compatibility issues when compiled
against an older version of the SDK.